# ARROW-10827: [Rust] Move concat from builders to a compute kernel and make it faster (2-6x)