@jorgecarleitao 2020-12-12
I have now migrated all tests to the `MutableArrayData`, which demonstrated a bug. The bug is fixed as part of this PR.